WindowsAppSDK fix

+++ platform/windows/publish.js
@@ -1,16 +1,36 @@
 import path from "node:path";
 import url from "node:url";
 import fs from "node:fs";
+import os from "node:os";
 import child_process from "node:child_process";
 import version from "../../version.js";
 
 /*
-* TO SELECT RELEASE OR BETA APP, RUN BEFORE PUBLISH
-* 
-* msstore init
-* 
-* IN platform/windows DIRECTORY
-*/
+ * TO SELECT RELEASE OR BETA APP, RUN BEFORE PUBLISH
+ *
+ * msstore init
+ *
+ * IN platform/windows DIRECTORY
+ */
+
+/*
+ *
+ * Fix for: https://github.com/microsoft/WindowsAppSDK/issues/5820
+ *
+ */
+const filePath = path.resolve(
+    os.homedir(),
+    ".nuget",
+    "packages",
+    "microsoft.windows.sdk.buildtools.msix",
+    "1.7.20250829.1",
+    "build",
+    "Microsoft.Windows.SDK.BuildTools.MSIX.Packaging.targets"
+);
+let contents = fs.readFileSync(filePath, { encoding: "utf-8" });
+contents.replace(/WinAppSdkSignAppxPackageBundle/g, "WinAppSdkSignAppxPackage");
+fs.writeFileSync(filePath, contents);
+// end fix
 
 const currentDirectory = path.dirname(url.fileURLToPath(import.meta.url));
 const rootDirectory = path.resolve(currentDirectory, "..", "..");
@@ -27,10 +47,13 @@ child_process.execSync(`npm run build ${noTSGO ? "-- --no-tsgo" : ""}`, {
 
 // build core
 
-child_process.execSync(`call ./windows.bat arm64 ${noTSGO ? "NO_TSGO=1" : ""}`, {
-    cwd: path.resolve(rootDirectory, "core", "build"),
-    stdio: "inherit"
-});
+child_process.execSync(
+    `call ./windows.bat arm64 ${noTSGO ? "NO_TSGO=1" : ""}`,
+    {
+        cwd: path.resolve(rootDirectory, "core", "build"),
+        stdio: "inherit"
+    }
+);
 
 child_process.execSync(`call ./windows.bat x64 ${noTSGO ? "NO_TSGO=1" : ""}`, {
     cwd: path.resolve(rootDirectory, "core", "build"),
@@ -44,17 +67,19 @@ child_process.execSync("call ./windows.bat copy", {
 
 // update version
 
-const versionStr = `${version.major}.${version.minor}.${version.build}.0`
+const versionStr = `${version.major}.${version.minor}.${version.build}.0`;
 const packageFile = path.resolve(currentDirectory, "Package.appxmanifest");
 let packageContent = fs.readFileSync(packageFile, { encoding: "utf-8" });
-packageContent = packageContent.replace(/\bVersion="\d+\.\d+\.\d+\.\d+"/g, `Version="${versionStr}"`)
+packageContent = packageContent.replace(
+    /\bVersion="\d+\.\d+\.\d+\.\d+"/g,
+    `Version="${versionStr}"`
+);
 fs.writeFileSync(packageFile, packageContent);
 
 // clean
 
 const appPackages = path.resolve(currentDirectory, "AppPackages");
-if(fs.existsSync(appPackages))
-    fs.rmSync(appPackages, { recursive: true })
+if (fs.existsSync(appPackages)) fs.rmSync(appPackages, { recursive: true });
 
 // msstore
